Separated state machine into modules, closes #52, closes #49
