+++ chain/chain/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-orga = { git = "https://github.com/nomic-io/orga", rev = "d7328eebc36c65479cfddcd15e86c1fc25291525", features = ["merk", "abci"] }
+orga = { git = "https://github.com/nomic-io/orga", rev = "2498df8f5e7871b61cd7c0812ee1b6b3a80569be", features = ["merk", "abci"] }
 nomic-primitives = { path = "../../core/primitives" }
 nomic-work = { path = "../../core/work" }
 nomic-bitcoin = { path = "../../core/bitcoin" }
