+++ core/bitcoin/Cargo.toml
@@ -8,6 +8,6 @@ edition = "2018"
 bitcoin = { version = "0.23.0", features = ["use-serde"] }
 bitcoincore-rpc = "0.9.0"
 serde = { version = "1.0.104", features = ["derive"] }
-orga = { git = "https://github.com/nomic-io/orga", rev = "d7328eebc36c65479cfddcd15e86c1fc25291525" }
+orga = { git = "https://github.com/nomic-io/orga", rev = "2498df8f5e7871b61cd7c0812ee1b6b3a80569be" }
