+++ core/primitives/Cargo.toml
@@ -10,5 +10,5 @@ serde = { version= "1.0.104", features = ["derive"] }
 failure = "=0.1.6"
 bincode = "1.2.1"
 sha2 = "0.8.1"
-orga = { git = "https://github.com/nomic-io/orga", rev = "d7328eebc36c65479cfddcd15e86c1fc25291525" }
+orga = { git = "https://github.com/nomic-io/orga", rev = "2498df8f5e7871b61cd7c0812ee1b6b3a80569be" }
 secp256k1 = "0.17.2"
