+++ core/signatory-set/Cargo.toml
@@ -13,7 +13,7 @@ bitcoin = "0.23.0"
 serde_json = "1.0.48"
 serde = "1.0.104"
 bincode = "1.2.1"
-orga = { git = "https://github.com/nomic-io/orga", rev = "d7328eebc36c65479cfddcd15e86c1fc25291525" }
+orga = { git = "https://github.com/nomic-io/orga", rev = "2498df8f5e7871b61cd7c0812ee1b6b3a80569be" }
 
 [dev-dependencies]
 secp256k1 = "0.17.2"
